+++ package.json
@@ -11,7 +11,7 @@
     "db:generate": "pnpm exec drizzle-kit generate:sqlite --config=drizzle.config.ts",
     "db:introspect": "pnpm exec drizzle-kit introspect:sqlite --config=drizzle.config.ts",
     "db:push": "pnpm exec drizzle-kit push:sqlite",
-    "db:seed": "tsx ./src/db/seed.ts --config=drizzle.config.ts"
+    "db:seed": "tsx ./src/db/seed/index.ts --config=drizzle.config.ts"
   },
   "dependencies": {
     "better-sqlite3": "^9.0.0",
