+++ src/db/data/comedy-movies.ts
@@ -1,4 +1,4 @@
-import { TmdbMovieItem } from '@/schemas/tmdbMovieItem';
+import { TmdbMovieItem } from '@/db/data/tmdbMovieItem.types';
 // 20 comedy movies taken from The Movie Database (TMDb)
 // https://api.themoviedb.org/3/genre/35/movies?api_key=00f3f32198696caff437631c007a7548&language=en-US&page=1
 // https://api.themoviedb.org/3/genre/35/movies?api_key=00f3f32198696caff437631c007a7548&language=en-US&page=2
